pyproject.toml
[project]
name = "RagBasedChatbot"
version = "0.0.1"
description = ""
authors = [{ name = "Deepak Sharma", email = "deepak.sharma@gmail.com" }]
readme = "README.md"
requires-python = "==3.11.9"

dependencies = [
  "streamlit>=1.28.0",
  "scikit-learn>=1.3.0",
  "numpy>=1.24.0",
  "tiktoken>=0.5.0",
  "python-dotenv>=1.0.0",
  "PyPDF2>=3.0.0",
  # Dependencies for normal RAG pipeline and vector store
  "faiss-cpu>=1.7.0",
  "sentence-transformers>=2.2.0",
  "chromadb>=0.4.0",
  "ollama>=0.1.7",
  "requests>=2.31.0"
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.9",
]

# Minimal dependencies for simple RAG (TF-IDF based)
simple = [
    "streamlit>=1.28.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0"
]

# Full dependencies for advanced RAG with embeddings and vector databases
full = [
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "ollama>=0.1.7",
    "requests>=2.31.0"
]

# Vector store only dependencies
vector-store = [
    "faiss-cpu>=1.7.0", 
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0"
]

# LLM only dependencies  
llm = [
    "ollama>=0.1.7",
    "requests>=2.31.0"
]

